+++ src/main/java/com/beca/concessionaria/dtos/responses/PostCarroResponse.java
@@ -4,8 +4,9 @@ import lombok.Data;
 
 @Data
 public class PostCarroResponse {
+
     private String marca;
     private String modelo;
     private String cor;
-    private Integer ano;
+    private Long ano;
 }
